util pkg refactoring Signed-off-by: Marco <llan@redhat.com>
