+++ operators/multiclusterobservability/pkg/certificates/cert_controller.go
@@ -24,7 +24,7 @@ import (
 	"github.com/open-cluster-management/addon-framework/pkg/addonmanager"
 	mcov1beta2 "github.com/open-cluster-management/multicluster-observability-operator/operators/multiclusterobservability/api/v1beta2"
 	"github.com/open-cluster-management/multicluster-observability-operator/operators/multiclusterobservability/pkg/config"
-	"github.com/open-cluster-management/multicluster-observability-operator/operators/multiclusterobservability/pkg/util"
+	"github.com/open-cluster-management/multicluster-observability-operator/operators/pkg/util"
 )
 
 const (
