+++ operators/multiclusterobservability/pkg/rendering/renderer_grafana.go
@@ -10,8 +10,8 @@ import (
 	"sigs.k8s.io/kustomize/v3/pkg/resource"
 
 	"github.com/open-cluster-management/multicluster-observability-operator/operators/multiclusterobservability/pkg/config"
-	"github.com/open-cluster-management/multicluster-observability-operator/operators/multiclusterobservability/pkg/util"
 	rendererutil "github.com/open-cluster-management/multicluster-observability-operator/operators/pkg/rendering"
+	"github.com/open-cluster-management/multicluster-observability-operator/operators/pkg/util"
 )
 
 func (r *MCORenderer) newGranfanaRenderer() {
