+++ operators/pkg/util/obj_compare_test.go
